refactor: Separate menubar instance initialization into a module
This commit refactors the code to extract the initialization of the menubar instance (mb) into a separate module called 'menubar.js'. By moving the menubar creation logic to its own module, I'm gonna to create customized shortcut function later.

Changes Made:
- Moved menubar instance creation logic from main.js to mbManager.js.
- Updated index.js to import and use the createMenubar function from menubar.js.
- Adjusted otherScript.js to import and use the getMenubarInstance function from mbManager.js.
